feat(blog): add SEO-friendly URLs for the blog posts
Add constraint with unique_for_date param to ensure that slugs are unique for the publication date.
Modify the URL pattern to use the publication date and slug for the post detail URL. Modify the
views: change the parameters of the post_detail view to match the new URL parameters. Modify
parameters of resolve function reverse() to corresponding modified URL pattern.